Move test for UnexpectedRowsExpectation into their own file
